requirements.txt
pdfplumber
pandas
streamlit
plotly
openpyxl
openai
requests
python-dotenv
pymupdf
google-genai



